fix: improve crm site tracking ui
